requirements.txt
aniso8601==9.0.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
flask-swagger-ui==4.11.1
greenlet==2.0.2
gunicorn==20.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mysql-connector-python==8.0.33
packaging==23.1
pluggy==1.0.0
psycopg2==2.9.6
pytest==7.3.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.3
Werkzeug==2.3.4
